Test reading all todos from the database.
This commit creates one parameterized test. This test tries to read all
todos from each of the testing json files, and checks to make sure that
the number of todos read matches the number of todos in the file.

To use parameterized tests, we added a dependency on
junit-jupiter-params to build.gradle

This commit addresses issue #9.