Add another test for reading all todos.
This test makes sure that the first todo read is also the first todo in
the database.

(This test makes the assumption that the database preserves the order of
its content. Given that FullUserListFromDB makes the same assumption, I
believe that this is a legitimate behavior to expect.)

This commit addresses issue #9.