Remove references to users from FullTodoListFromDB
The file `FullTodoListFromDB.java` was mostly written by
copying-and-pasting from `FullUserListFromDB.java` (which is a sin, I
know). In the process, some references to users were not changed.

This commit fixes that. It rewrites the javadoc comment on
FullTodoListFromDB, and changes all references of Database to
TodoDatabase.

This commit addresses issue #9.