Fix erroneous tests
There were a several things wrong with the code that tests limiting the number of todos.

In LimitTodosFromDB.java:
1. In the testListTodosWithLimit method, we were passing the query parameter "category" rather than "limit".
2. The number we were passing for the "limit" query parameter was being passed as an int when in needed to be a String.
3. Several identifiers had erroneous names: categoryTodos should have been limitedTodos, and 
categoriesAndCountsTodosParams should have been limitsAndCountsTodosParams.

These are all copy-and-paste bugs. (And textbook examples of why I shouldn't be copying-and-pasting.)

in TodoControllerSpec:
In the GET_to_request_limit_2_todos and GET_with_illegal_limit_throws_error methods, we were passing the query parameter "status" rather than "limit".

This is yet another copy-and-paste bug.