Test that negative limits are treated as zero
This test currently fails, because negative limits currently throw an IllegalArgumentException inside the TodoDatabase.filterTodosWithLimit method. (On the client side, this manifests itself as a server error, which isn't good.)

I'm deciding that negative limits should be treated as zero because that limits the amount of annoying error messages that the user has to see. A case could be made, though, that the server should still throw an error, and we should just clean up the error message. (Similar to what we're doing with non-numeric inputs.)