Test sorting by invalid fields
This test is currently failing, because of the edgiest of edge cases: if the database has no todos in it—or, for that matter, if it has one todo—it doesn't matter if we're sorting by a field that doesn't exist, because *the sort function never tries to access that field.* No errors are thrown. The server is happy to sort an empty array by field names like "heffalump", because it never actually looks at the field "heffalump".